lecture20: IIFE - Immediately Invoked Function Expression
The function in the code works as a module. The external parenthesis work as a wrapper,
making the function to work as an expression, and therefore, can be used in the
code, such as being invoked.

The invokation is accomplished by the parentheis right after its declaration.

In fact, we can pass a value to it, and it will receive it right after being invoked.

An IIFE creates a scope in the code, and can only be accessed inside the IIFE, with the
same intent we can use a module.